used ros_components | created package hardware and robot
