+++ rosbot/CMakeLists.txt
@@ -1,0 +1,6 @@
+cmake_minimum_required(VERSION 3.8)
+project(rosbot)
+
+find_package(ament_cmake REQUIRED)
+
+ament_package()
