+++ rosbot/rosbot_hardware.repos
@@ -1,0 +1,9 @@
+repositories:
+  rosbot_hardware_interfaces:
+    type: git
+    url: https://github.com/husarion/rosbot_hardware_interfaces.git
+    version: main
+  ros_components_description:
+    type: git
+    url: https://github.com/husarion/ros_components_description.git
+    version: ros2_update_to_guide
