+++ rosbot_bringup/launch/rosbot_bringup.launch.py
@@ -1,113 +1,29 @@
 from launch import LaunchDescription
-from launch.substitutions import Command
 from launch_ros.actions import Node
-from launch.substitutions import Command, PathJoinSubstitution
+from launch.substitutions import  PathJoinSubstitution
 from ament_index_python.packages import get_package_share_directory
-from launch_ros.substitutions import FindPackageShare
-from launch.event_handlers import OnProcessExit
 from launch.actions import IncludeLaunchDescription
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from launch.substitutions import PathJoinSubstitution
-from launch.actions import RegisterEventHandler
 
 
 def generate_launch_description():
-    rosbot_description = get_package_share_directory("rosbot_description")
+    rosbot_hardware = get_package_share_directory("rosbot_hardware")
     rosbot_bringup = get_package_share_directory("rosbot_bringup")
 
-    rosbot_description_launch = IncludeLaunchDescription(
+    rosbot_hardware_launch = IncludeLaunchDescription(
         PythonLaunchDescriptionSource(
             PathJoinSubstitution(
                 [
-                    rosbot_description,
+                    rosbot_hardware,
                     "launch",
-                    "rosbot_description.launch.py",
+                    "controller.launch.py",
                 ]
             )
         )
     )
 
-    xacro_file = PathJoinSubstitution([rosbot_description, "urdf", "rosbot.urdf.xacro"])
     ekf_config = PathJoinSubstitution([rosbot_bringup, "config", "ekf.yaml"])
-
-    robot_description = {
-        "robot_description": Command(["xacro --verbosity 0 ", xacro_file])
-    }
-
-    robot_controllers = PathJoinSubstitution(
-        [
-            FindPackageShare("rosbot_bringup"),
-            "config",
-            "rosbot_controllers.yaml",
-        ]
-    )
-
-    control_node = Node(
-        package="controller_manager",
-        executable="ros2_control_node",
-        parameters=[robot_description, robot_controllers],
-        remappings=[
-            ("/imu_sensor_node/imu", "/_imu/data_raw"),
-            ("~/motors_cmd", "/_motors_cmd"),
-            ("~/motors_response", "/_motors_response"),
-            ("/rosbot_base_controller/cmd_vel_unstamped", "/cmd_vel"),
-        ],
-    )
-
-    joint_state_broadcaster_spawner = Node(
-        package="controller_manager",
-        executable="spawner",
-        arguments=[
-            "joint_state_broadcaster",
-            "--controller-manager",
-            "/controller_manager",
-            "--controller-manager-timeout",
-            "120",
-        ],
-    )
-
-    robot_controller_spawner = Node(
-        package="controller_manager",
-        executable="spawner",
-        arguments=[
-            "rosbot_base_controller",
-            "--controller-manager",
-            "/controller_manager",
-            "--controller-manager-timeout",
-            "120",
-        ],
-    )
-
-    # Delay start of robot_controller after joint_state_broadcaster
-    delay_robot_controller_spawner_after_joint_state_broadcaster_spawner = (
-        RegisterEventHandler(
-            event_handler=OnProcessExit(
-                target_action=joint_state_broadcaster_spawner,
-                on_exit=[robot_controller_spawner],
-            )
-        )
-    )
-
-    imu_broadcaster_spawner = Node(
-        package="controller_manager",
-        executable="spawner",
-        arguments=[
-            "imu_broadcaster",
-            "--controller-manager",
-            "/controller_manager",
-            "--controller-manager-timeout",
-            "120",
-        ],
-    )
-
-    # Delay start of imu_broadcaster after robot_controller
-    # when spawning without delay ros2_control_node sometimes crashed
-    delay_imu_broadcaster_spawner_after_robot_controller_spawner = RegisterEventHandler(
-        event_handler=OnProcessExit(
-            target_action=robot_controller_spawner,
-            on_exit=[imu_broadcaster_spawner],
-        )
-    )
 
     robot_localization_node = Node(
         package="robot_localization",
@@ -118,12 +34,8 @@ def generate_launch_description():
     )
 
     actions = [
-        rosbot_description_launch,
-        control_node,
+        rosbot_hardware_launch,
         robot_localization_node,
-        joint_state_broadcaster_spawner,
-        delay_robot_controller_spawner_after_joint_state_broadcaster_spawner,
-        delay_imu_broadcaster_spawner_after_robot_controller_spawner,
     ]
 
     return LaunchDescription(actions)
